# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "this is paro api.",
    "version": "1.0.0",
    "title": "Swagger parto"
  },
  "servers": [
    {
      "url": "https://api.parto.app"
    },
    {
      "url": "http://localhost:2216"
    },
    {
      "url": "https://dev.parto.app"
    }
  ],
...
  "paths": {
...
    "/auth/logIn/{lang}": {
      "post": {
        "summary": "login(sign in / sign up) user with phone.",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "imei is not requier and can be not in body",
                "required": [
                  "phone"
                ],
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "description": "set one of the \"Lady\" or \"Husband\" or \"Teenager\""
                  },
                  "imei": {
                    "type": "string"
                  }
                },
                "example": {
                  "phone:": "09199698086",
                  "type": "Lady"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "x-auth-token": {
                "required": true,
                "description": "token set to response header",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "description": "user id"
                        },
                        "type": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          }
        }
      }
    },
...
    "/auth/verifyCode": {
      "post": {
        "summary": "Returns survey question.",
        "tags": [
          "user"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "send sms successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "description": "verify code",
                      "properties": {
                        "message": {
                          "type": "string"
                        },
                        "code": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "invalid input (phone number/email)"
          },
          "418": {
            "description": "Your account credit is not enough"
          },
          "502": {
            "description": "server error to send email"
          }
        }
      }
    },
    "/auth/signIn/{lang}": {
      "post": {
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          }
        ],
        "summary": "Creates a new user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "description": "user id"
                        },
                        "userName": {
                          "type": "string",
                          "description": "user id"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "phone exist / invalid input"
          }
        }
      }
    },
...
    "/auth/partnerVerifyCode/{userId}/{lang}": {
      "post": {
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "use in husband version and set partner to database and check husband code is true or no",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "partnerCode": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "invalid entery when userId is wrong or partnerId(in the code) is wrong or partnerCode mathematically is wrong"
          }
        }
      }
    },
    "/auth/partnerVerifyCode/{userId}/{lang} ": {
      "get": {
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "use in husband version and set partner to database and check husband code is true or no",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "partnerCode": {
                          "type": "string",
                          "description": "code that gnerate for partner"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "invalid entery when userId is wrong or partnerId(in the code) is wrong or partnerCode mathematically is wrong"
          }
        }
      }
    },
...
  }
}